Todo's can successfully be removed and maintain their appropriate index/indentifiers. The 'index-indentifiers' of the todo elements and ToDo objects in the todos and ToDos arrays respectively will be used to reference/target specifc todo elements and ToDo objects in order to make changes or remove them.
Everytime a todo or ToDo is created it is given an identifier that correlates to it's position in the todo and ToDo arrays. Because ToDo Objects and todo Elements are created together and added to their respective arrays in the same operation, their indexes are correlated. Whenever a todo element is removed the removeToDo() function is invoked and updates the identifier's of the elements in the two arays, so that they are re-correlated and accurate to their index in their respective array.