Projects can now be added to the nav section
The 'Add Project' button in the header has a 'click' event on it that will call the addProject function in the add-porject.js module. This module will generate a Project element using the ProjectComponent from the components directory, and will append it to the Nav section for display. Simaultaneously a project object is created using the projectFactory function from the project-object.js module. Now we can create Projects - storing the information about them in their own object, adding them to an array to be referenced and indexed and creating an element to represent them in the UI.