Remove project bug fixed:
Projects can now be successfully removed from the UI and from our array storing those project-objects. The bug came from a missed line of code in the remove-project.js file, in the logic directory, in the for loop that updates each index value of the items storing the Project UI elements (projectElements). I missed adding a line that updates the index property of the current project element iteration to the value of the index (i). With this change, the app now can successfully remove projects from the UI and it's corresponding object.