Basic Logic for navigating Projects and assigning todos to their corresponding project's 'todos' array, created.
In the module project-nav.js in 'logic' directory, the 'current state' of our application will be tracked, meaning this module will be used to track the project that a user is currently active/opened on. This will be used to determine UI rendering of Todos from the current project's todos and rendering and retrieving of project information when necessary.

Variable, 'userState', is an object that stores the current state that will be later based on UI events + logic. Function setCurrentProject(event) will be used to set the current state based on a UI event by the user (clicking into a project). currentProject() returns the Project that corresponds to the current state, by iterating through the allProjects array (where all projects are stored) and returning the first object (project) that matches the condition (project's name === current state), this will be used to determine what Project new Todos will be saved to (Project's todos array property).