Form now loads and successfully creates a Project and stores it in the 'allprojects' array.
First: event listener on 'Add Project' title in the project Nav for a 'click', when clicked will render the projectForm() component and form for a new project renders. Second: When the 'submit' event for the form happens it calls the addProject function from the add-project module. This will then generate a new project and pass the user's input from the form to the projectFactory function imported from the project-object.js module. Third: The projectFactory function will generate a new project with the passed input and push that object to the allProjects array, to be later referenced and used.