Todos can now be successfully created and pushed to the 'currentProject()' and it's todos array property.
First: event listener on 'Add Todo' title in the todos-section for a 'click', when clicked will render the todoForm() component and form and a new todo renders. Second: When the 'submit' event for the form happens it calls the addTodo function from the add-todo module.js. This will then generate a new todo by passing the user's input from the form to the todoFactory function imported from the todo-object.js module. Third: The todoFactory function will generate a new todo and push that object to the 'todos' array property of the current Project that it is being added to, to be later referenced and used.