Projects can be deleted
Projects can now be removed both as objects in the allProjects array and from the UI. An eventlistener is attached to the 'x' (delete) button, with the handler removeProject(event) from the remove-project.js module. This handler will take the event and obtain it's 'deleteName' property (a property assigned to the 'x' element that will have the same value as the name of the project and the same 'elementName' of the UI element containing the Project title in the nav). deleteName will then be used to locate the index of the project with the same name, that's correlated to the UI element and will be removed from the allProjects array. After, deleteProjectElements(projectName) is called. This function will locate the UI element based on the argument passed (the value of the deleteName property of the correlated 'x' element) and will remove it from the DOM.