getProjectIndex(projectName) created - Used to find the index of a project in the allProjects array.
Found myself rewriting code to find the index of a project in the allProjects array, using the findIndex() method on the allProjects array. So I created a function to reuse this functionality, located in the project-nav.js module. This function takes an argument that is the value of the project's name property you want to find, and using the findIndex() array method on the allProjects array, we locate the first project name property that matches that of the arugment passed, and return the index number.