Todo items can be removed from the UI and their respective project's todos array.
Todo objects can now be removed through the use of the removeTodo(event) in the remove-todo.js module. This function will find the todo object based on the event target's elementName and project properties (these properties are assigned upon rendering via renderTodos() and renderAllTodos() and are given values of the todo's name and the project's name values respectively). These properties are then used to find the index of the associated project in the allProjects array, access it's todos property array (where todos associated with that project are held) and then the index of the todo who's name property value matches that of the elementName of the event target. After the todo is spliced from it's project's todos array, removeTodoElement() from TodoSection.js module is invoked and removes the element who's event was called on.