Form validation. Added form validation for both Project and Todo forms. This will check to see if the names submitted for the Todo/Project objects matches any of the other projects/todos that exist. This way we can avoid duplicating names, which will allow us to keep selecting and finding projects and todos based on their specific names to identify.
This functionality is achieved through the use of isDuplicatedProjectName() and isDuplicatedTodoName() in the form-validation.js module. Both utilize the JS marray method, some(). This method will return a true or false if the condition is true at any point iterating through the array. In our case we're comparing the user's inputted title value against each name of all the projects in the allProjects array, and in the case of todo names - we're comparing this title value against all of the todo names in that projects todos array. Using an if/else on the 'submit' event listeners for the forms we can then send an alert to tell the user what they did wrong and how to correct it.