Todos now take date input and description input.
Users can now enter in a date and description for the todos they create. Using the date-fns library I was able to add a date input field in the 'todoForm' and pass it to the todoFactory() in the todo-object.js module, and give the newly created todo objects a method called formatDate(), that will parse and format the date using the parse and format methods from the date-fns library. This way date properties from a todo can be converted into a more readable string format, that can also be used to render this in the UI also.