Inline editing renders to UI
The elements necessary to render the inline editing to the UI, allowing users to change the name of their projects 'in-place'. The renderInlineEditing(projectName) function in the EditProject.js module, selects the nav-item element associated with the 'click' event (on the edit-icon), deletes all the elements within it and then renders the elements for the inline/in-place editing. The inline editing consists of a <span> where users can input their new project title into and a <div> container that holds a 'cancel' and 'confirm' icon. Currently, both of these icons just revert the Nav back to it's original state, but eventually 'confirm' will send the textContent of the <span> to change the associated project's name.