Todo stylings for different priorities
Todo's rendered in the UI now have a different styling applied to them based on their set 'priority' by the user. This was done using some simple CSS and classes. High priority todos have red checkboxes with bolded text and mid priorty todos have orange checkboxes with normal font-weight text. Upon rendering the todos with the renderTodos() in the TodoSection.js module the checkbox and todo-title elements are given a specifc class based on priority, these are: 'checkbox-mid' and 'checkbox-high' for checkboxes  and 'title-mid' and 'title-high' for todo titles.